+++ api/app/Http/Resources/BandMemberSetupSummaryResource.php
@@ -16,6 +16,7 @@ class BandMemberSetupSummaryResource extends JsonResource
             'instrument_name'   => $this->instrument?->name,
             'shared_monitor_id' => $this->shared_monitor_id,
             'name'              => $this->name,
+            'is_default'        => (bool) $this->is_default,
             'signal_chain_type' => $this->signal_chain_type,
             'input_count'       => count($this->inputs ?? []),
             'updated_at'        => $this->updated_at,